refactor: restructure the typescript project
