+++ .gitignore
@@ -7,3 +7,4 @@ coverage
 .electrom
 dist
 .eslintcache
+lib
