+++ example/renderer/index.tsx
@@ -1,16 +1,21 @@
 import { useState } from 'react';
 import ReactDOM from 'react-dom';
 import classnames from 'classnames';
-import StatusBoard from '../../src/StatusBoard';
-import PerfBoard from '../../src/PerfBoard';
+import { PreloadElectron } from '../../src/common/window';
+import { StatusBoard, PerfBoard, EVENT_DATA_CHANNEL_NAME, EVENT_ACTION_CHANNEL_NAME } from '../../src/renderer';
 
 import 'antd/dist/antd.css';
 
 import styles from './index.module.less';
 
+declare global {
+  interface Window {
+    electron: PreloadElectron;
+  }
+}
+
 const container = document.querySelector('#app');
 
-const params = new URLSearchParams(location.search);
 const { ipcRenderer, shell } = window.electron;
 
 const App = () => {
@@ -18,8 +23,8 @@ const App = () => {
   return (
     <>
       <StatusBoard
-        eventDataChannelName={params.get('EVENT_DATA_CHANNEL_NAME') || ''}
-        eventActionChannelName={params.get('EVENT_ACTION_CHANNEL_NAME') || ''}
+        eventDataChannelName={EVENT_DATA_CHANNEL_NAME}
+        eventActionChannelName={EVENT_ACTION_CHANNEL_NAME}
         ipcRenderer={ipcRenderer}
         shell={shell}
       />
