--- example/renderer/preload.d.ts
@@ -1,16 +1,0 @@
-import { Shell } from 'electron';
-
-declare global {
-  interface Window {
-    electron: {
-      ipcRenderer: {
-        send: (channel: string, ...args: any[]) => void;
-        on: (channel: string, listener: any) => void;
-        removeListener: (channel: string, listener: any) => void;
-      };
-      shell: Shell;
-    };
-  }
-}
-
-export {};
