--- lib/electrom.ts
@@ -1,2 +1,0 @@
-export { PerfTracing } from './perf/tracing';
-export * from './monitor';
