+++ src/common/interface.ts
@@ -1,17 +1,17 @@
 export interface ProcessInfo {
-  cmd: string;
+  cmd?: string;
+  load?: number;
+  sandboxed?: boolean;
   cpu: {
     percentCPUUsage: number;
     idleWakeupsPerSecond: number;
   };
   creationTime: number;
-  load: number;
   memory: {
     workingSetSize: number;
     peakWorkingSetSize: number;
   };
   pid: number;
-  sandboxed: boolean;
   type: string;
   webContentInfo?: {
     id: number;
