--- src/components/WebContentsDetail/index.module.less
@@ -1,28 +1,0 @@
-.wrapper {
-  height: 200px;
-
-  .info {
-    margin-left: 8px;
-  }
-
-  .detail {
-    margin: 8px;
-    font-size: 12px;
-    overflow-wrap: break-word;
-  }
-
-  .buttons {
-    display: flex;
-    justify-content: flex-end;
-    padding: 8px;
-
-    button {
-      margin-left: 8px;
-
-      &:first-child {
-        margin-left: 0;
-      }
-    }
-  }
-
-}
