+++ src/main/index.ts
@@ -1,0 +1,2 @@
+export * from './perf/tracing';
+export * from './monitor';
