+++ src/renderer/StatusBoard.module.less
