+++ src/renderer/components/BottomPanel/index.module.less
@@ -3,6 +3,7 @@
   bottom: 0;
   left: 0;
   padding: 16px;
+  height: 250px;
   display: flex;
   width: 100%;
   flex-direction: column;
