+++ src/renderer/components/ProcessDetail/index.module.less
@@ -1,0 +1,33 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+
+  .detail {
+    margin: 8px;
+    font-size: 12px;
+    flex: 1;
+    max-height: 200px;
+    overflow: auto;
+  }
+
+  .control {
+    position: absolute;
+    bottom: 0;
+    right: 0;
+    padding: 10px;
+
+    .buttons {
+      display: flex;
+      justify-content: flex-end;
+      padding: 8px;
+
+      button {
+        margin-left: 8px;
+
+        &:first-child {
+          margin-left: 0;
+        }
+      }
+    }
+  }
+}
