+++ src/renderer/components/WebContentsDetail/index.module.less
@@ -1,0 +1,35 @@
+.wrapper {
+  height: 200px;
+
+  .info {
+    margin-left: 8px;
+  }
+
+  .detail {
+    margin: 8px;
+    font-size: 12px;
+    overflow-wrap: break-word;
+  }
+
+  .control {
+    position: absolute;
+    bottom: 0;
+    right: 0;
+    padding: 10px;
+
+    .buttons {
+      display: flex;
+      justify-content: flex-end;
+      padding: 8px;
+
+      button {
+        margin-left: 8px;
+
+        &:first-child {
+          margin-left: 0;
+        }
+      }
+    }
+  }
+
+}
