+++ src/renderer/index.tsx
@@ -1,0 +1,3 @@
+export * from './PerfBoard';
+export * from './StatusBoard';
+export * from '../common/constants';
